Fix reusing ESP partition
